+++ demo/sdk_demo_v2/SETTINGS_virtualBG_workflow.h
@@ -1,6 +1,7 @@
 #pragma once
 #include "sdk_util.h"
 #include "stdafx.h"
+#include "sdk_demo_app_common.h"
 #include "setting_service_interface.h"
 
 class CSDKVirtualBGUIEvent
@@ -24,10 +25,11 @@ public:
 	DWORD GetBGReplaceColor();
 
 	ZOOM_SDK_NAMESPACE::SDKError SetUsingGreenScreen(bool bUse);
-	ZOOM_SDK_NAMESPACE::SDKError AddBGImage(const wchar_t* file_path);
-	ZOOM_SDK_NAMESPACE::SDKError RemoveBGImage(ZOOM_SDK_NAMESPACE::IVirtualBGImageInfo* pRemoveImage);
+	ZOOM_SDK_NAMESPACE::SDKError AddBGImage(const wchar_t* file_path, bkItemType itemType);
+	ZOOM_SDK_NAMESPACE::SDKError RemoveBGImage(ZOOM_SDK_NAMESPACE::IVirtualBGImageInfo* pRemoveImage, bkItemType itemType);
 	ZOOM_SDK_NAMESPACE::IList<ZOOM_SDK_NAMESPACE::IVirtualBGImageInfo* >* GetBGImageList();
-	ZOOM_SDK_NAMESPACE::SDKError UseBGImage(ZOOM_SDK_NAMESPACE::IVirtualBGImageInfo* pImage);
+	ZOOM_SDK_NAMESPACE::IList<ZOOM_SDK_NAMESPACE::IVirtualBGImageInfo* >* GetBGVideoList();
+	ZOOM_SDK_NAMESPACE::SDKError UseBGImage(ZOOM_SDK_NAMESPACE::IVirtualBGImageInfo* pImage, bkItemType itemType);
 	ZOOM_SDK_NAMESPACE::SDKError BeginSelectReplaceVBColor();
 	ZOOM_SDK_NAMESPACE::SDKError SetVideoPreviewParentWnd(HWND hParentWnd, RECT rc /* = _SDK_TEST_VIDEO_INIT_RECT */);
 	ZOOM_SDK_NAMESPACE::SDKError TestVideoStartPreview(const wchar_t* deviceID /* = NULL */);