+++ h/zoom_sdk_def.h
@@ -54,7 +54,8 @@ enum SDKError
 	SDKERR_MEETING_VIEWTYPE_PARAMETER_IS_WRONG, ///<Incorrect ViewType parameters.
 	SDKERR_MEETING_ANNOTATION_IS_OFF, ///<Annotation is disabled.
 	SDKERR_SETTING_OS_DONT_SUPPORT, ///<Current OS doesn't support the setting.
-	SDKERR_EMAIL_LOGIN_IS_DISABLED, //Email login is disable
+	SDKERR_EMAIL_LOGIN_IS_DISABLED, ///<Email login is disable
+	SDKERR_HARDWARE_NOT_MEET_FOR_VB, ///<Computer doesn't meet the minimum requirements to use virtual background feature.
 };
 
 /*! \enum SDK_LANGUAGE_ID
@@ -171,6 +172,35 @@ enum ZoomSDKVideoRenderMode
 	ZoomSDKVideoRenderMode_GDI,
 };
 
+enum ZoomSDKRenderPostProcessing
+{
+	ZoomSDKRenderPostProcessing_None = 0,
+	ZoomSDKRenderPostProcessing_Auto,
+	ZoomSDKRenderPostProcessing_Enable,
+	ZoomSDKRenderPostProcessing_Disable,
+};
+
+enum ZoomSDKVideoCaptureMethod
+{
+	ZoomSDKVideoCaptureMethod_None = 0,
+	ZoomSDKVideoCaptureMethod_Auto,
+	ZoomSDKVideoCaptureMethod_DirectSHow,
+	ZoomSDKVideoCaptureMethod_MediaFoundation,
+};
+
+typedef struct tagZoomSDKRenderOptions
+{
+	ZoomSDKVideoRenderMode    videoRenderMode;
+	ZoomSDKRenderPostProcessing renderPostProcessing;
+	ZoomSDKVideoCaptureMethod videoCaptureMethod;
+	tagZoomSDKRenderOptions()
+	{
+		videoRenderMode = ZoomSDKVideoRenderMode_None;
+		renderPostProcessing = ZoomSDKRenderPostProcessing_Auto;
+		videoCaptureMethod = ZoomSDKVideoCaptureMethod_Auto;
+	}
+}ZoomSDKRenderOptions;
+
 typedef struct tagRawDataOptions
 {
 	bool enableRawdataIntermediateMode; ///<false -- YUV420data, true -- intermediate data
@@ -185,7 +215,7 @@ typedef struct tagRawDataOptions
 		audioRawdataMemoryMode = ZoomSDKRawDataMemoryModeStack;
 	}
 }RawDataOptions;
-		
+
 /*! \struct tagInitParam
     \brief Initialize the SDK Parameter.
     Here are more detailed structural descriptions.
@@ -205,7 +235,7 @@ typedef struct tagInitParam
 	ConfigurableOptions obConfigOpts;///<The configuration options of the SDK.
 	SDK_APP_Locale locale;
 	bool permonitor_awareness_mode;
-	ZoomSDKVideoRenderMode    videoRenderMode;
+	ZoomSDKRenderOptions renderOpts;
 	RawDataOptions rawdataOpts;
 	tagInitParam()
 	{
@@ -221,7 +251,6 @@ typedef struct tagInitParam
 		uiLogFileSize = 5;
 		locale = SDK_APP_Locale_Default;
 		permonitor_awareness_mode = true;
-		videoRenderMode = ZoomSDKVideoRenderMode_None;
 	}
 }InitParam;